Moved the System.Json extension methods to a new class JsonAssertionExtensions
