+++ Src/FluentAssertions/AssertionExtensions.cs
@@ -23,11 +23,6 @@ using NotNullAttribute = System.Diagnostics.CodeAnalysis.NotNullAttribute;
 using FluentAssertions.Events;
 #endif
 
-#if NET6_0_OR_GREATER
-using System.Linq;
-using System.Text.Json.Nodes;
-#endif
-
 namespace FluentAssertions;
 
 /// <summary>
@@ -907,26 +902,6 @@ public static class AssertionExtensions
     public static TaskCompletionSourceAssertions Should(this TaskCompletionSource tcs)
     {
         return new TaskCompletionSourceAssertions(tcs, AssertionChain.GetOrCreate());
-    }
-
-    /// <summary>
-    /// Returns an object that provides various assertion APIs that act on a <see cref="JsonNode"/>.
-    /// </summary>
-    [Pure]
-    public static JsonNodeAssertions<T> Should<T>([NotNull] this T jsonNode)
-        where T : JsonNode
-    {
-        return new JsonNodeAssertions<T>(jsonNode, AssertionChain.GetOrCreate());
-    }
-
-    /// <summary>
-    /// Return an object that provides various assertion APIs that treat a <see cref="JsonArray"/>
-    /// as a collection of <see cref="JsonNode"/>s.
-    /// </summary>
-    [Pure]
-    public static GenericCollectionAssertions<JsonNode> Should([NotNull] this JsonArray jsonArray)
-    {
-        return new GenericCollectionAssertions<JsonNode>(jsonArray?.ToArray(), AssertionChain.GetOrCreate());
     }
 
 #endif
